+++ src/main/java/com/auctionwebsite/service/impl/UserDetailsImpl.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 public class UserDetailsImpl implements UserDetails {
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private int id;
 
     private String username;
 
@@ -26,7 +26,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Long id, String email, String password, String username,
+    public UserDetailsImpl(int id, String email, String password, String username,
                            Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
